read the description
so like when you run it, the program does something. pretty amazing isnt it its still not a game unless you concider looking at a chesspiece in the console to be a game in which case you're in luck because boy do I have something for you. If you're a normal person however you can sit there and be amazed by how unfinished the code is like i have probably 30 things to do in the project and i forgot half and im going to do them as they come up which isnt the greatest philosophy but eventually the game will work also i have literally no idea how i will implement 90% of the necisary features but i guess ill cross that bridge when i get there also to the 0 people reading this bye bye and also the code has gotten significantly worse to look at like it woud probably make joey puke and that is like so awful i would never want that to happen to joey and yet if he saw this he would so yeah i guess my new goal is to keep this code away from joey bye bye now